Revert "TRA-16658 - Sélecteur de code Bâle (#4719)"
This reverts commit 36ad88618a0be40aeb14ce78739ce0da6079dfe4.
