+++ front/src/form/registry/common/WasteCodeSelector.scss
@@ -24,7 +24,3 @@
     }
   }
 }
-
-.search-btn-info-label {
-  margin-bottom: 56px;
-}
