+++ front/src/form/registry/incomingTexs/shape.ts
@@ -15,7 +15,6 @@ import {
   optionalString
 } from "../builder/validation";
 import { Address } from "../common/Address";
-import { BaleCodeTreeSelector } from "../common/BaleCodeTreeSelector";
 import { CompanySelector } from "../common/CompanySelector";
 import { EcoOrganismes } from "../common/EcoOrganismes";
 import { ParcelsVisualizer } from "../common/ParcelsVisualizer/ParcelsVisualizer";
@@ -90,19 +89,15 @@ export const incomingTexsFormShape: FormShape = [
         style: { className: "fr-col-md-10" }
       },
       {
-        Component: BaleCodeTreeSelector,
-        props: {
-          name: "wasteCodeBale",
-          label: Labels.wasteCodeBale,
-          infoLabel: InfoLabels.wasteCodeBale,
-          required: false
-        },
-        shape: "custom",
-        names: ["wasteCodeBale"],
+        name: "wasteCodeBale",
+        shape: "generic",
+        label: Labels.wasteCodeBale,
+        infoLabel: InfoLabels.wasteCodeBale,
         validation: {
           wasteCodeBale: optionalString
         },
-        style: { parentClassName: "fr-grid-row--bottom tw-relative" }
+        type: "text",
+        style: { className: "fr-col-md-4" }
       },
       {
         name: "wastePop",
