+++ front/src/form/registry/outgoingWaste/shape.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 import { FormShape } from "../builder/types";
 import { ReportFor } from "../common/ReportFor";
-import { BaleCodeTreeSelector } from "../common/BaleCodeTreeSelector";
 import { WasteCodeSelector } from "../common/WasteCodeSelector";
 import { WeightSelector } from "../common/WeightSelector";
 import {
@@ -85,19 +84,15 @@ export const outgoingWasteFormShape: FormShape = [
         style: { className: "fr-col-12 fr-col-md-10" }
       },
       {
-        Component: BaleCodeTreeSelector,
-        props: {
-          name: "wasteCodeBale",
-          label: Labels.wasteCodeBale,
-          infoLabel: InfoLabels.wasteCodeBale,
-          required: false
-        },
-        shape: "custom",
-        names: ["wasteCodeBale"],
+        name: "wasteCodeBale",
+        shape: "generic",
+        label: Labels.wasteCodeBale,
+        infoLabel: InfoLabels.wasteCodeBale,
         validation: {
           wasteCodeBale: optionalString
         },
-        style: { parentClassName: "fr-grid-row--bottom tw-relative" }
+        type: "text",
+        style: { className: "fr-col-md-4" }
       },
       {
         name: "wastePop",
